Drop `as any` casts; fix stale view imports and vitest collection
Atoms (eliminate 'as any' at call sites):
- atoms.ts: replace per-call-site `as any` casts with one targeted
  `as PrimitiveAtom<T>` at each atom declaration (seekEventAtom,
  streamValueAtom). `atom<T | null>(null)` resolves to a read-only
  Atom under Jotai's overloads — the bare `null` arg matches the
  read-fn overload first — so writable shape has to be reasserted
  somewhere. Doing it once at declaration is type-safe and confined.
- use-playback-engine.ts: drop the call-site `store.set(seekEventAtom
  as any, ...)` cast.
- stream-base.ts: drop the call-site `store.set(streamValueAtom(id)
  as any, ...)` cast.

Pre-existing fixes surfaced by tsc / vitest:
- use-create-timeline.ts: `return undefined` instead of bare `return`
  (TS7030 under noImplicitReturns).
- views/PlaybackElements.tsx, views/Timeline.tsx, views/TimelineExamples.tsx:
  fix imports from `../lib/...` to `../lib/timeline/...` after the
  timeline files moved into a subdirectory; prefix unused `range`
  param and drop unused `getFrameNumber` destructure.
- timeline/timeline-name.ts (new): pure getTimelineNameFromSampleAndGroupId
  helper, no @fiftyone/state dep. utils.ts pulls from here so that
  utils.test.ts (and any other test importing utils.ts) doesn't
  transitively load @fiftyone/relay's graphql template tag, which
  fails at module load without the babel transform.
  use-default-timeline-name.ts re-exports the helper so external
  consumers keep their existing import path.
