+++ app/packages/playback/src/lib/timeline/use-default-timeline-name.ts
@@ -2,21 +2,10 @@ import * as fos from "@fiftyone/state";
 import { useCallback, useMemo } from "react";
 import { useRecoilValue } from "recoil";
 import { GLOBAL_TIMELINE_ID } from "../constants";
+import { getTimelineNameFromSampleAndGroupId } from "./timeline-name";
 
-export const getTimelineNameFromSampleAndGroupId = (
-  sampleId?: string | null,
-  groupId?: string | null
-) => {
-  if (!sampleId && !groupId) {
-    return GLOBAL_TIMELINE_ID;
-  }
-
-  if (groupId) {
-    return `timeline-${groupId}`;
-  }
-
-  return `timeline-${sampleId}`;
-};
+// Re-exported so existing consumers keep their import path.
+export { getTimelineNameFromSampleAndGroupId };
 
 /**
  * This hook gives access to the default timeline name based on the current context.
